+++ build-scripts/for-linux-armhf/build-on-linux.sh
@@ -7,7 +7,8 @@
 
 set -e
 
-source $(dirname $0)/../set-ver-prms.sh "$1" "$2"
+DIR=$(dirname $0)
+source $DIR/../set-ver-prms.sh "$1" "$2"
 
 if [[ -n "$3" ]]; then
 	SRC_DIR=$3
@@ -40,7 +41,12 @@ GO_ARM_PRMS="-DCMAKE_SYSTEM_NAME=Linux \
   -DCPACK_DEBIAN_PACKAGE_ARCHITECTURE=armhf \
   -DCPACK_RPM_PACKAGE_ARCHITECTURE=armhf \
   -DIMPORT_EXECUTABLES=../build-tools/ImportExecutables.cmake"
-GO_PRMS="-DCMAKE_BUILD_TYPE=Release $CMAKE_VERSION_PRMS -DCMAKE_PACKAGE_SUFFIX=$PACKAGE_SUFFIX"
+
+GO_PRMS="-DCMAKE_BUILD_TYPE=Release \
+  $CMAKE_VERSION_PRMS \
+  -DCMAKE_PACKAGE_SUFFIX=$PACKAGE_SUFFIX \
+  $($DIR/../for-linux/cmake-prm-yaml-cpp.bash armhf)"
+
 echo "cmake -G \"Unix Makefiles\" $GO_PRMS . $SRC_DIR"
 cmake -G "Unix Makefiles" $GO_PRMS $GO_ARM_PRMS . $SRC_DIR
 make -k $PARALLEL_PRMS VERBOSE=1 package