+++ src/grandorgue/model/GOManual.cpp
@@ -538,9 +538,6 @@ void GOManual::Reset() {
   for (unsigned j = 0; j < GetDivisionalCount(); j++)
     GetDivisional(j)->Display(false);
 
-  if (GetStopCount() == 1 && !GetStop(0)->IsDisplayed())
-    return;
-
   for (unsigned j = 0; j < GetStopCount(); j++)
     GetStop(j)->Reset();
 }